Add animations of block groups falling
