+++ module.txt
@@ -6,6 +6,5 @@
     "description" : "Detects groups of blocks disconnected from the ground to let them fall",
     "dependencies" : [
         { "id" : "Health", "minVersion" : "1.0.2" }
-    ],
-    "serverSideOnly" : true
+    ]
 }
