+++ src/main/java/org/terasology/fallingblocks/calculation/node/UnloadedNode.java
@@ -1,14 +1,14 @@
-// Copyright 2020 The Terasology Foundation
+// Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-package org.terasology.fallingblocks.node;
+package org.terasology.fallingblocks.calculation.node;
 
 import org.joml.Vector3i;
-import org.terasology.fallingblocks.Chain;
-import org.terasology.fallingblocks.FullChain;
-import org.terasology.fallingblocks.Pair;
-import org.terasology.fallingblocks.Tree;
-import org.terasology.fallingblocks.TreeUtils;
+import org.terasology.fallingblocks.calculation.Chain;
+import org.terasology.fallingblocks.calculation.FullChain;
+import org.terasology.fallingblocks.calculation.Pair;
+import org.terasology.fallingblocks.calculation.Tree;
+import org.terasology.fallingblocks.calculation.TreeUtils;
 
 import java.util.HashSet;
 import java.util.Set;
