+++ src/main/java/org/terasology/fallingblocks/calculation/updates/LoadUpdate.java
@@ -1,13 +1,13 @@
-// Copyright 2020 The Terasology Foundation
+// Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-package org.terasology.fallingblocks.updates;
+package org.terasology.fallingblocks.calculation.updates;
 
 import org.joml.Vector3i;
-import org.terasology.fallingblocks.Chain;
-import org.terasology.fallingblocks.Tree;
-import org.terasology.fallingblocks.TreeUtils;
-import org.terasology.fallingblocks.node.Node;
+import org.terasology.fallingblocks.calculation.Chain;
+import org.terasology.fallingblocks.calculation.Tree;
+import org.terasology.fallingblocks.calculation.TreeUtils;
+import org.terasology.fallingblocks.calculation.node.Node;
 
 import java.util.Set;
