+++ src/main/java/org/terasology/fallingblocks/calculation/updates/RemovalUpdate.java
@@ -1,11 +1,11 @@
-// Copyright 2020 The Terasology Foundation
+// Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-package org.terasology.fallingblocks.updates;
+package org.terasology.fallingblocks.calculation.updates;
 
 import org.joml.Vector3i;
-import org.terasology.fallingblocks.Chain;
-import org.terasology.fallingblocks.Tree;
+import org.terasology.fallingblocks.calculation.Chain;
+import org.terasology.fallingblocks.calculation.Tree;
 
 import java.util.Set;
