+++ src/main/java/org/terasology/fallingblocks/calculation/updates/Update.java
@@ -1,12 +1,12 @@
-// Copyright 2020 The Terasology Foundation
+// Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-package org.terasology.fallingblocks.updates;
+package org.terasology.fallingblocks.calculation.updates;
 
 import java.util.*;
 
-import org.terasology.fallingblocks.Chain;
-import org.terasology.fallingblocks.Tree;
+import org.terasology.fallingblocks.calculation.Chain;
+import org.terasology.fallingblocks.calculation.Tree;
 
 /**
  * A message from the main thread to the UpdateThread to modify the octree somehow.
