+++ src/main/java/org/terasology/fallingblocks/calculation/updates/ValidateUpdate.java
@@ -1,10 +1,10 @@
-// Copyright 2020 The Terasology Foundation
+// Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-package org.terasology.fallingblocks.updates;
+package org.terasology.fallingblocks.calculation.updates;
 
-import org.terasology.fallingblocks.Chain;
-import org.terasology.fallingblocks.Tree;
+import org.terasology.fallingblocks.calculation.Chain;
+import org.terasology.fallingblocks.calculation.Tree;
 
 import java.util.Collections;
 import java.util.Set;
